cmake: Fix how compile definition is added
The compile definition ENET_ENHANCEDBUFFERDESCRIPTOR_MODE has been
added by modifying the build environment of the 'zephyr' library. But
modifying this library is not supported.

The intention was to modify the global build environment, so we change
to use the zephyr_ API instead of the zephyr_library_ API as this API
will behave as intended.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>