lpc54114: Remove offsets_short.h inclusion.
This commit removes unnecessary inclusion of offsets_short.h in the
LPC54114 start-up code.

The offsets_short.h is a kernel private header and must not be included
by non-kernel components. The kernel and arch private header
directories will be removed from the compiler include path in the near
future.

For more details, Refer to the issue zephyrproject-rtos/zephyr#20119.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>