zephyr/module.yml: Update to use cmake-ext/kconfig-ext
Move to using cmake-ext/kconfig-ext which means we'll look in the
zephyr repo for cmake and kconfig files.  This allows us to put
zephyr specific integrations with zephyr itself.

We'll have the zephyr side cmake include the module cmake so this
should effectively be transparent at this point.  (we'd do the same
for kconfig, but there are no kconfig files in this repo).

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>