driver: enet: Convert to compatible with 64-bit platforms
Fix the size of 'buffer' in RX and TX buffer descriptors,
The buffer pointer has fixed size of 4 Bytes in buffer
descriptors, but the current type of 'buffer' is 'uint8_t *',
it isn't compatible with 64-bit platforms, so change its type
to uint32_t. And add assert to verify the given data buffers
are located in 32-bit adddress range.

And fix build warnings on 64-bit platforms

Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>