s32: s32z27: remove suffix on number of SIUL2 channels
The external interrupts shim driver uses the
SIUL2_ICU_IP_NUM_OF_CHANNELS macro together with listify, therefore the
integer literal suffix must be removed in order for the macros to
concatenate correctly. This way we can keep the shim driver generic
enough when other SoC is supported and there is no need to redefine the
macro there.

Signed-off-by: Manuel Arguelles <manuel.arguelles@nxp.com>