mcux: mcux-sdk: fix calculation of Flexcomm USART BRG value in 32K mode
Fix calculation of the USART BRG value when using the flexcomm in 32KHz
clock mode. This value should be one less, or the resulting baud rate
will be half what the user requests. This issue is already fixed in
the USART_Enable32kMode() function, but still existed in the
USART_Init() function.

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>