devices: MIMX8QM6: Add mapped versions of clock functions
With this commit, we can make use of Zephyr's device_map() functionality
which maps a physical address during a driver's initialization. This way
we can avoid using mmu_regions.c

Since an identity mapping may not be emploied, it's necessary to pass
the LPCG's virtual (or physical for devices with no MMU ) address as an argument.

Signed-off-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>