s32: drivers: s32k344_adc: remove suffix on number of ADC instances
The ADC shim driver uses the ADC_INSTANCE_COUNT macro together with
listify, therefore the integer literal suffix must be removed in order
for the macros to concatenate correctly. This way we can keep the shim
driver generic enough when other SoC is supported and there is no need
to redefine the macro there.

Signed-off-by: Cong Nguyen Huu <cong.nguyenhuu@nxp.com>