s32: s32k344: remove suffix from GMAC instance count
The GMAC shim driver uses FEATURE_GMAC_NUM_INSTANCES as argument of
LISTIFY, therefore the integer literal suffix must be removed in order
for the macros to concatenate correctly. This way we can keep the shim
driver generic enough when other SoC is supported and there is no need
to redefine the macro there.

Signed-off-by: Manuel Argüelles <manuel.arguelles@nxp.com>