mcux: scripts: update soc header script to handle cfg reg type correctly
Update SOC header script for IMX parts to handle configuration register
type correctly. RT11xx SOCs have multiple types of configuration
registers for pins, and this needs to be encoded in the devicetree for
the MUX options. Fix the header generation script to correctly evaluate
the config type based on which bitfields are in the register, instead of
based on the pin name (as this pattern is not valid for all GPIO_DISP
pins)

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>