drivers: irqsteer: Add revision 2 of the irqsteer driver
This commit introduces revision 2 of the IRQ_STEER driver.
The need for this revision stems from the fact that the
previous one was built around the M cores, thus making it
unusable for the other cores (e.g: the DSP core). As such,
this revision attempts to make the driver more generic to
make it usable on all cores. Introducing a new revision is
necessary so as to not break applications using the old driver.

Signed-off-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>