MIMX8QM6: drivers: fsl_clock.c: Report NULL frequency for unsupported core
Currently, if the core is not supported the compilation will
fail because of the "#error" statement in CLOCK_GetCoreSysClkFreq.
To overcome this, instead of causing a compilation error, the
function will report a NULL frequency as if the IPC call failed.
This way, we can avoid having to define macros for each of the
cores just to make CLOCK_GetCoreSysClkFreq() happy.

Signed-off-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>