s32: soc: s32z27: integrate with zephyr
Remove 'u' suffix from macros so that they
can be used with LISTIFY in shim drivers.

Set the number of UART LINFlexD instances configured
that bases on the number of devicetree UART nodes enabled.

Set the number of SPI instances configured that bases
on the number of devicetree SPI nodes enabled.
Enable SPI slave support bases on CONFIG_SPI_SLAVE.

Wrap the defined macros of each MRU instance base on
devicetree MRU node so that it is built when node enabled.

Signed-off-by: Cong Nguyen Huu <cong.nguyenhuu@nxp.com>