s32: mcux: use the SoC name to set the device name
Switch to use `CONFIG_SOC` to define the device name and CPU name
needed to build MCUX drivers and device sources.
The SoC part number for S32 now makes reference to the actual
part number of device as defined in datasheets and should't be
used to define the CPU name.

Signed-off-by: Manuel Argüelles <manuel.arguelles@nxp.com>