mcux: hal_nxp.cmake: Allow NULL FIFO watermark for Zephyr DAI SAI driver
The Zephyr DAI SAI driver may want to set a NULL FIFO watermark
in which case the NXP HAL SAI driver needs to allow it as it's
a valid value. As such, this commit sets the MCUX_SDK_SAI_ALLOW_NULL_FIFO_WATERMARK
macro to 1 which will allow the NULL FIFO watermark.

Signed-off-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>