mcux: hal_nxp: update path for scfw_api
With HWMv2, when trying to load the scfw_api for 8QM or
8QXP I get:
```
include could not find requested file:
    driver_scfw_api
```
So update the path for scfw_api based on MCUX_DEVICE_PATH.

For ADSP we don't have separate devices (i.e. MIMX8QX6_dsp
or MIMX8QM6_dsp), because will duplicate a lot of code.
For these we use the core devices (i.e. MIMX8QX6 or MIMX8QM6).

Signed-off-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>
Signed-off-by: Iuliana Prodan <iuliana.prodan@nxp.com>