S32K142: Added DTS and clock bindings
These additions should be preferrably pushed to zephyr upstream but
I am not a mantainer of the NXP SoCs. I am commiting this for
the sake of convenience. If NXP is interested in supporting the
S32K142 feel free to use this commit and push these files to
zephyr upstream.

Signed-off-by: Victor Chavez <vchavezb@protonmail.com>