s32: s32z: patch rtd 2.0.0 to work with zephyr
Remove 'u' suffix from macros so that they can be used
with LISTIFY.

Netc shim driver uses 'Netc_Eth_Ip_MSIX_Rx' function. Need
to remove 'static inline' keywords from this function so
that it can be used as an extern function outside of the
file declaration.

Set the number of UART LINFlexD instances configured that
bases on the number of devicetree UART nodes enabled.

Set the number of SPI instances configured that bases on
the number of devicetree SPI nodes enabled.

Enable SPI slave support bases on CONFIG_SPI_SLAVE.

Wrap the defined macros of each MRU instance base on
devicetree MRU node so that it is built when node enabled.

Signed-off-by: Ha Duong Quang <ha.duongquang@nxp.com>