mcux: mcux-sdk: drivers: cache64: remove invalid assert in GetInstance
CACHE64_GetInstanceByAddr() asserts when an address is passed in that is
not managed by the CACHE64 controller. This is incorrect behavior, as
the CACHE64 driver checks to see if the returned instance number is out
of range when using this function, so we should not assert here.

Remove the assert so that cache invalidation functions can be used with
addresses not the CACHE64 controller's range.

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>