devices: MIMX8ML8: fsl_clock: change root clock ID for SAI3
According to the TRM, the parent of the SAI3 bus clock is not
SAI3_ROOT, but AUDIO_AHB_CLK_ROOT. As such, the currently used
root parent ID (77) is wrong.

Furthermore, CLOCK_EnableClock() and CLOCK_DisableClock() also
attempt to ungate/gate the root parent of the AUDIOMIX clock.
Since AUDIO_AHB_CLK_ROOT is the parent of multiple SAI bus clocks,
allowing it to be gated during CLOCK_DisableClock() is not safe.

As such, set the parent of SAI3 bus clock to 0xFFFF. This way,
CLOCK_DisableClock() and CLOCK_EnableClock() will not attempt
to gate/ungate it. This means that gating/ungating AUDIO_AHB_CLK_ROOT
will have to be performed explicitly.

Signed-off-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>