zephyr: blobs: update MCXW71 BLE controller to MCUXSDK 24.12.00 release
This new version of the BLE controller is aligned with the one for
MCXW72. It contains a fix which consists of returning an HCI command
complete event when setting the BD address.
We need both platforms to be aligned on this behavior to correctly
implement our zephyr hci_nxp driver.

Signed-off-by: Axel Le Bourhis <axel.lebourhis@nxp.com>