devices: MIMXRT685S: Rename hifi4 to dsp
Rename MIMXRT685S/hifi4/ to MIMXRT685S/dsp/. Remove reference to
${SdkRootDirPath}/arch/xtensa/configuration/common.cmake.

This is done because the SDK originally uses the "_dsp" core suffix to
identify the HiFi 4 domain. This same suffix (CONFIG_MCUX_CORE_SUFFIX)
is used both in the core subdirectory and the core-specific header files
(such as the MIMXRT685S_dsp.h). As the header files come from the SDK
and the core subdirectory is an addition for this Zephyr HAL, it's
easier to use the somewhat inaccurate suffix "_dsp".

The common.cmake file doesn't even exist in this tree nor in the SDK.

Signed-off-by: Vit Stanicek <vit.stanicek@nxp.com>