devices: MIMXRT685S: Rename hifi4 to dsp
Rename MIMXRT685S/hifi4/ to MIMXRT685S/dsp/.

This is done because the SDK originally uses the "_dsp" core suffix to
identify the HiFi 4 domain. This same suffix (CONFIG_MCUX_CORE_SUFFIX)
is used both in the core subdirectory and the core-specific header files
(such as the MIMXRT685S_dsp.h). As the header files come from the SDK
and the core subdirectory is an addition for this Zephyr HAL, it's
easier to use the somewhat ambiguous suffix "_dsp".

Signed-off-by: Vit Stanicek <vit.stanicek@nxp.com>