hal_nxp: mcux: README: Add notice about SAI and CAAM driver versions
During the hal_nxp 25.09.00 update, the SAI driver was
upgraded to the latest SDK version, which introduced
issues on MCXN947 in Zephyr I2S-related test cases.
To avoid regressions, add a notice in the README advising
against updating the SAI driver as part of hal_nxp updates.

In the same update, the CAAM driver was intentionally
kept at an older version because the latest SDK driver
introduced issues on RT1160/RT1170 in Zephyr networking
test cases. Add a notice in the README advising against
updating the CAAM driver in future hal_nxp updates.

Signed-off-by: Zhaoxiang Jin <Zhaoxiang.Jin_1@nxp.com>