mcux: wifi_nxp: Add compiler flag for the wake_timer
Add new compiler flag CONFIG_WAKE_TIMER_ENABLE for the wake_timer. The
duration of this timer is 5 seconds and it is used to block low power
mode for 5 seconds after device wakes up from low power mode.
User can choose to enable this timer or not by controlling this compiler
flag.

Signed-off-by: Hui Bai <hui.bai@nxp.com>